Moved login helpers to utilities.
