+++ src/js/user-profile/components/UserDataSection.jsx
@@ -7,7 +7,7 @@ import AlertBox from '@department-of-veterans-affairs/jean-pants/AlertBox';
 import LoadingIndicator from '@department-of-veterans-affairs/jean-pants/LoadingIndicator';
 import Modal from '@department-of-veterans-affairs/jean-pants/Modal';
 import recordEvent from '../../../platform/monitoring/record-event';
-import { mfa } from '../../../platform/site-wide/login/utils/helpers';
+import { mfa } from '../../../platform/utilities/authentication';
 import AcceptTermsPrompt from '../../common/components/AcceptTermsPrompt';
 import { fetchLatestTerms, acceptTerms } from '../actions';
 import PersonalizationBetaInvite from './PersonalizationBetaInvite';
